requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pinecone-client>=2.2.4
openai>=1.3.0
langchain>=0.0.350
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.10
pydantic>=2.4.2
sqlalchemy>=2.0.23
asyncpg>=0.29.0
python-multipart>=0.0.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2
numpy>=1.24.0
tenacity>=8.0.0 